🔧 Moves Shared Project Reference into Dependent Projects
This moves the reference to `FastGraph.Tests.CommonInternals` from the `Directory.Build.props` into the individual project files that use it. This aims to avert a transient build failure where the related projects sometimes suddenly fail to be able to build due to an apparent inability to resolve standard library types.